==> notebooks/0- Cleaning Catalogue DF.ipynb <==
# Processing the exported dataframe from Magento database#
     With details of the products and their category associated
       

Both dataframes are in a folder register into .gitignore

import pandas as pd

df = pd.read_csv('../data2/raw/logiscenter_com.csv',delimiter=';',low_memory=False)

df.head()

df['attribute_set'].value_counts()

Dropping columns that are not needed

df.drop(columns=['entity_id','modelo','manufacturer','descatalogado','variant','category_ids','category_names*'], inplace=True)

df.head(2)

# Exporting df to a csv on a .gitignore folder #
(because it has businessl details)

df.to_csv('../final_data/processed/catalogue_clean.csv', index = False)
