==> notebooks/0- Cleaning Sales DF.ipynb <==
## 0. Importing and reading df with sales

import pandas as pd

df = pd.read_csv('../data2/raw/clientes_sku_ventas.csv')

df

## 1. Cleaning ##

- Cleaning VAT column. The one will use to merge with the companies enriched dataset

df['vat'] = df['vat'].str.replace('ES','')
df['vat'] = df['vat'].str.strip()

- Cleaning name.1 column. The one will use to merge with the categorized sku dataset

df['name.1'] = df['name.1'].str.strip()
df['name.1'] = df['name.1'].str.replace('PO:','')

- creating a 'sku' column

df['sku'] = df['name.1'].str.extract('\[(.+)\]', expand=True)

- Recounting number of null values in new 'sku' column

df[df['sku'].isnull()]['name.1'].value_counts()

Reindexing neede to do a loop for

- Manually assigning 'sku' values to top seller products with incorrect "name.1" codification

for i in range(len(df)):
    if df.loc[i,'name.1'] == 'DS9208-SR KIT USB EMEA U21':
        df.loc[i,'sku'] = 'DS9208-SR4NNU21Z'
    elif df.loc[i,'name.1'] == 'VOYAGER 9520 USB BLACK KIT MASKED STAND + HARD MOUNT':
        df.loc[i,'sku'] = 'MK9520-32A38'
    elif df.loc[i,'name.1'] == 'GK420D DT 203DPI USB SER PARAL LEL R2.0':
        df.loc[i,'sku'] = 'GK42-200120-000'
    elif df.loc[i,'name.1'] == 'VOYAGER MS9520 KIT USB BLACK MASKED STAND + HARD MOUNT':
        df.loc[i,'sku'] = 'MK9520-32A38'
    elif df.loc[i,'name.1'] == 'CN3 WIFI BTH  2D IMGR NUM 128/256 WM6,1':
        df.loc[i,'sku'] = 'CN3BNH80000E400'
    elif df.loc[i,'name.1'] == 'GK420T TT 203DPI USB SER PARAL LEL R2.0':
        df.loc[i,'sku'] = 'GK42-100120-000'
    elif df.loc[i,'name.1'] == 'SERVICIOS DE REPARACION':
        df.loc[i,'sku'] = 'Repair'
    elif df.loc[i,'name.1'] == 'LS2208 KIT USB BLACK STAND N':
        df.loc[i,'sku'] = 'LS2208-7AZU0100S'
    elif df.loc[i,'name.1'] == 'Shipping and Handling':
        df.loc[i,'sku'] = 'Shipping and Handling'
    elif df.loc[i,'name.1'] == 'ECLIPSE 5145 USB BLACK KIT NO STAND CG CANNOT BE DEACTIVA':
        df.loc[i,'sku'] = 'MK5145-71A38-EU'
    elif df.loc[i,'name.1'] == 'CABLE POWER   2M EU     TO IEC':
        df.loc[i,'sku'] = 'AK5015'
    elif df.loc[i,'name.1'] == 'VOYAGER 1200G BLACK USB KIT RIGID STAND / TYPE-A USB CABLE':
        df.loc[i,'sku'] = '1200g-2USB-1F02'
    elif df.loc[i,'name.1'] == 'VOYAGER-GS 1250G BLACK SCANNER':
        df.loc[i,'sku'] = '1250G-2KBW-1'
    elif df.loc[i,'name.1'] == 'SUMINISTROS VARIOS':
        df.loc[i,'sku'] = 'unknown'
    elif df.loc[i,'name.1'] == 'MISCELLANEOUS':
        df.loc[i,'sku'] = 'unknown'
    elif df.loc[i,'name.1'] == 'QUANTUM-E IS3480 RS232 EU KIT EU PSU + POWERLINK CABLE':
        df.loc[i,'sku'] = 'MK3480-30B41'
    elif df.loc[i,'name.1'] == 'RBN IX SERIES YMCKO  ZXP3':
        df.loc[i,'sku'] = '800033-344'
    elif df.loc[i,'name.1'] == 'QUICKSCAN MOBILE M2130 BLK KIT 433 MHZ USB W/BASE-CBL':
        df.loc[i,'sku'] = 'QD2130-BKK11S'
    elif df.loc[i,'name.1'] == 'MC75A6 HSDPA, 802.11a/b/g, 1D Cam Numeric WM6':
        df.loc[i,'sku'] = 'MC75A6-P0CSWQRA9WR'

- References not classified yet

df[df['sku'].isnull()]['name.1'].value_counts()

- For 'sku' affected by the expanded column that are not readable, change them to common category

for i in range(len(df)):
    if df.loc[i,'sku'] == 'SHIP':
        df.loc[i,'sku'] = 'Shipping and Handling'
    if df.loc[i,'sku'] == 'TR.EXPRESS':
        df.loc[i,'sku'] = 'Shipping and Handling'
    if df.loc[i,'sku'] == 'SSREP':
        df.loc[i,'sku'] = 'Repair'
    if df.loc[i,'sku'] == 'DIF-PREC':
        df.loc[i,'sku'] = 'Price-offer'
    if df.loc[i,'sku'] == 'DISCOUNT.PP':
        df.loc[i,'sku'] = 'Price-offer'
    if df.loc[i,'sku'] == 'DESC.PP':
        df.loc[i,'sku'] = 'Price-offer'
    if df.loc[i,'sku'] == 'EUR-MISC':
        df.loc[i,'sku'] = 'unknown'

df['sku'].value_counts()

df.head()

Deleting columns that are not needed

df.drop(columns=['name','name.1'], inplace = True)

df.head()

# Exporting df to a csv on a .gitignore folder #
(because it has client personal details)

df.to_csv('../final_data/processed/sales_clean.csv', index = False)
