==> notebooks/1- Merge sales-catalogue.ipynb <==
# In order to categorized product we will merge two dataframes.#
        1- df with sales by sku and client
        2- df with catalogue exported from magento with sku and category


Both dataframes are in a folder register into .gitignore

import pandas as pd

df = pd.read_csv('../final_data/processed/merged_clean.csv', index_col = 0)

df.head(2)

dfs.head(2)

dfs.shape

## Will Merge using the column 'sku' from both df

final = pd.merge(dfs,dfc, on='sku', how = 'left')

final

Will drop all the registers without a valid value on 'attribute_set'

final['attribute_set'].isnull().sum()

final[final['attribute_set'].isnull()]['sku'].value_counts()

final = final[final['attribute_set'].notnull()]

final.shape

final

Will not count lines with negative values

final = final[final['price_subtotal'] > 0]

final.shape

Will drop 'sku' column as no longer needed in order to anonymize business information

final = final.drop(columns=('sku'))

Will aggregate values by vat and attribute set

final

aggresults = final.groupby(['id','vat','attribute_set']).agg({'price_subtotal':['sum'], 'quantity':['sum']})

aggresults

aggresults

aggresults = aggresults.reset_index(level=[0,1])
aggresults = aggresults.reset_index(level=[0])

aggresults.columns = [col[1] for col in aggresults.columns]

aggresults.columns = ['category','id', 'vat', 'value', 'quantity']

aggresults

df = aggresults[['id','vat','category','value','quantity']]

df.drop_duplicates()

Will use only sales with very concrete categories dropping "Other"

df = df[df['category'] != 'Other']

df

df.to_csv('../final_data/processed/merged_clean.csv')
