==> notebooks/3- Data engineering einforma .ipynb <==
import pandas as pd

companies = pd.read_csv('../final_data/processed/final_cifs_enrich.csv', index_col = 0)

companies = companies.drop(columns = ['index', 'Unnamed: 0.1'])

companies['provincia'] = companies['Localidad'].str.extract('\((.+)\)', expand=True)

companies = companies.drop(columns = ['Localidad', 'Actividad Informa', 'Objeto Social','Forma Juridica'])

companies['provincia'] = companies['provincia'].astype(str)
companies['CNAE'] = companies['CNAE'].astype(str)


Lets do some LabelEncoder

from sklearn.preprocessing import LabelEncoder

labelencoder = LabelEncoder()

companies['provincia_encoded'] = labelencoder.fit_transform(companies['provincia'])
companies['cnae_encoded'] = labelencoder.fit_transform(companies['CNAE'])


companies

companies.to_csv('../final_data/processed/final_companies.csv')
