==> notebooks/4- Final DATAFRAME.ipynb <==
import pandas as pd

df = pd.read_csv('../data2/processed/merged_clean.csv', index_col = 0)

df['value'] = df['value'].astype(float)
df['quantity'] = df['quantity'].astype(int)

total_values = df.groupby('vat')['value'].sum()
total_quantity = df.groupby('vat')['quantity'].sum()


df.head()

df1 = pd.merge(df, total_values, on='vat')


df1.head()

aggresults = pd.merge(df1, total_quantity, on='vat')

aggresults.rename(columns={'value_x':'value','quantity_x':'quantity','value_y':'total_value', 'quantity_y':'total_quantity'}, inplace = True)

aggresults['category'].value_counts()

cat_list = list(aggresults['category'].unique())
cat_list

for x in cat_list:
    aggresults[x+'_value'] = aggresults[aggresults['category'] == x]['value']
    aggresults[x+'_qty'] =  aggresults[aggresults['category'] == x]['quantity']

 aggresults.head()

aggresults.drop(columns=['category','value','quantity'], inplace = True)

aggresults.fillna(0)

columnas = list(aggresults.columns)
columnas

dataframe = aggresults.groupby(['id','vat','total_value','total_quantity'])['Wristband_value',
 'Wristband_quantity',
 'RFID_value',
 'RFID_quantity',
 'Tickets_value',
 'Tickets_quantity',
 'CardPrinter_value',
 'CardPrinter_quantity',
 'TouchScreen_value',
 'TouchScreen_quantity',
 'Card_value',
 'Card_quantity',
 'CardRibbon_value',
 'CardRibbon_quantity',
 'Mobile_value',
 'Mobile_quantity',
 'Scanner_value',
 'Scanner_quantity',
 'LabelPrinter_value',
 'LabelPrinter_quantity',
 'Label_value',
 'Label_quantity',
 'Ribbon_value',
 'Ribbon_quantity',
 'TicketPrinter_value',
 'TicketPrinter_quantity',
 'Printhead_value',
 'Printhead_quantity',
 'Accessories_value',
 'Accessories_quantity'].sum().reset_index()


dataframe['vat'].value_counts()

# We will merge with all the information scrapped about the companies

companies = pd.read_csv('../data2/processed/final_companies.csv', index_col = 0)

companies

compa2 = companies[['vat','provincia_encoded','cnae_encoded']]

df = pd.merge(dataframe, compa2, on = 'vat')

df.head()

df = df.drop_duplicates() 

df.fillna(0)

df.to_csv('../data2/processed/DATAFRAME.csv')
